AUDIT_ID更改为Long类型

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/saas/shop/mp/MpAuthShopVo.java
@@ -111,7 +111,7 @@ public class MpAuthShopVo {
 	/**
 	 * 提交审核ID
 	 */
-	private Integer auditId;
+	private Long auditId;
 	/**
 	 * 提交审核状态
 	 */
